Refactor project tree
