+++ frontend/src/lib/init_font_awesome.ts
