+++ frontend/src/views/routes.ts
@@ -4,7 +4,7 @@ const routes: Array<RouteRecordRaw> = [
     {
         path: '/',
         name: 'Home',
-        component: () => import('@/components/Home.vue'),
+        component: () => import('@/views/Home.vue'),
     },
 ];
